requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.4
click==8.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.9.0
nba_api==1.5.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.5
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0
pytz==2025.2
referencing==0.37.0
requests==2.32.3
rich==13.9.4
rpds-py==0.28.0
scikit-learn==1.5.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.39.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
lxml==6.0.2
firebase-admin==6.5.0
python-dotenv==1.0.0
